+++ citywok_ms/__init__.py
@@ -1,8 +1,10 @@
-from flask import Flask
+from flask import Flask, request, current_app
 from flask_sqlalchemy import SQLAlchemy
+from flask_babel import Babel
 from citywok_ms.config import Config
 
 db = SQLAlchemy()
+babel = Babel()
 
 
 def create_app(config_class=Config):
@@ -12,6 +14,7 @@ def create_app(config_class=Config):
 
     # init extensions
     db.init_app(app)
+    babel.init_app(app)
 
     with app.app_context():
         # imports
@@ -21,3 +24,8 @@ def create_app(config_class=Config):
         # TODO:
 
         return app
+
+
+@babel.localeselector
+def get_local():
+    return request.accept_languages.best_match(current_app.config['LANGUAGES'])