+++ citywok_ms/utils.py
@@ -1,0 +1,8 @@
+SEX = [('M', 'Male'),
+       ('F', 'Female')]
+
+ID = [('passport', 'Passport'),
+      ('id_card', 'ID Card'),
+      ('green_card', 'Permanent Resident Card'),
+      ('residence_permit', 'Residence Permit'),
+      ('other', 'Other')]